+++ quantum_debugger/algorithms/__init__.py
@@ -188,6 +188,7 @@ from .imaginary_time import imaginary_time_evolution
 from .adiabatic import adiabatic_evolution
 from .krylov import krylov_spectrum, krylov_ground_energy
 from .loschmidt import loschmidt_echo, rate_function, quench_dynamics
+from .otoc import otoc, scrambling_time
 from .spectroscopy import unitary_eigenphase, hermitian_eigenvalue
 
 __all__ = [
@@ -380,6 +381,8 @@ __all__ = [
     "loschmidt_echo",
     "rate_function",
     "quench_dynamics",
+    "otoc",
+    "scrambling_time",
     # Quantum spectroscopy
     "unitary_eigenphase",
     "hermitian_eigenvalue",